+++ packages/react-native/src/wrap.tsx
@@ -5,7 +5,12 @@ import { HotUpdaterError } from "./error";
 import { useEventCallback } from "./hooks/useEventCallback";
 import { reload, updateBundle } from "./native";
 import type { RunUpdateProcessResponse } from "./runUpdateProcess";
-import { type HotUpdaterState, useHotUpdaterStore } from "./store";
+import { useHotUpdaterStore } from "./store";
+
+type UpdateStatus =
+  | "CHECK_FOR_UPDATE"
+  | "UPDATING"
+  | "UPDATE_PROCESS_COMPLETED";
 
 export interface HotUpdaterConfig extends CheckForUpdateConfig {
   /**
@@ -29,7 +34,10 @@ export interface HotUpdaterConfig extends CheckForUpdateConfig {
    *
    * If not defined, the bundle will download in the background without blocking the screen.
    */
-  fallbackComponent?: React.FC<Pick<HotUpdaterState, "progress">>;
+  fallbackComponent?: React.FC<{
+    status: Exclude<UpdateStatus, "UPDATE_PROCESS_COMPLETED">;
+    progress: number;
+  }>;
   onError?: (error: HotUpdaterError) => void;
   onProgress?: (progress: number) => void;
   /**
@@ -54,13 +62,12 @@ export function wrap<P>(
   return (WrappedComponent) => {
     const HotUpdaterHOC: React.FC<P> = () => {
       const { progress } = useHotUpdaterStore();
-      const [status, setStatus] = useState<
-        "IDLE" | "CHECK_FOR_UPDATE" | "UPDATING" | "UPDATE_PROCESS_COMPLETED"
-      >("IDLE");
+      const [updateStatus, setUpdateStatus] =
+        useState<UpdateStatus>("CHECK_FOR_UPDATE");
 
       const initHotUpdater = useEventCallback(async () => {
         try {
-          setStatus("CHECK_FOR_UPDATE");
+          setUpdateStatus("CHECK_FOR_UPDATE");
           const updateInfo = await checkForUpdate({
             source: restConfig.source,
             requestHeaders: restConfig.requestHeaders,
@@ -69,12 +76,23 @@ export function wrap<P>(
             restConfig.onUpdateProcessCompleted?.({
               status: "UP_TO_DATE",
             });
-            setStatus("UPDATE_PROCESS_COMPLETED");
+            setUpdateStatus("UPDATE_PROCESS_COMPLETED");
             return;
           }
 
-          setStatus("UPDATING");
+          if (updateInfo.shouldForceUpdate === false) {
+            void updateBundle(updateInfo.id, updateInfo.fileUrl);
+            restConfig.onUpdateProcessCompleted?.({
+              id: updateInfo.id,
+              status: updateInfo.status,
+              shouldForceUpdate: updateInfo.shouldForceUpdate,
+            });
+            setUpdateStatus("UPDATE_PROCESS_COMPLETED");
+            return;
+          }
 
+          // Force Update Scenario
+          setUpdateStatus("UPDATING");
           const isSuccess = await updateBundle(
             updateInfo.id,
             updateInfo.fileUrl,
@@ -85,7 +103,7 @@ export function wrap<P>(
             );
           }
 
-          if (updateInfo.shouldForceUpdate && reloadOnForceUpdate) {
+          if (reloadOnForceUpdate) {
             reload();
           }
 
@@ -94,12 +112,12 @@ export function wrap<P>(
             status: updateInfo.status,
             shouldForceUpdate: updateInfo.shouldForceUpdate,
           });
-          setStatus("UPDATE_PROCESS_COMPLETED");
+          setUpdateStatus("UPDATE_PROCESS_COMPLETED");
         } catch (error) {
           if (error instanceof HotUpdaterError) {
             restConfig.onError?.(error);
           }
-          setStatus("UPDATE_PROCESS_COMPLETED");
+          setUpdateStatus("UPDATE_PROCESS_COMPLETED");
           throw error;
         }
       });
@@ -114,10 +132,10 @@ export function wrap<P>(
 
       if (
         restConfig.fallbackComponent &&
-        status !== "UPDATE_PROCESS_COMPLETED"
+        updateStatus !== "UPDATE_PROCESS_COMPLETED"
       ) {
         const Fallback = restConfig.fallbackComponent;
-        return <Fallback progress={progress} />;
+        return <Fallback progress={progress} status={updateStatus} />;
       }
 
       return <WrappedComponent />;